[BACKLOG-2892]  Changes to allow parameter values to be honored in dataservice SQL.
For example, "WHERE PARAMETER('foo') = 'bar'" should set the 'foo' parameter to 'bar'.
This functionality was broken at some point when a check was added for non-existent
fieldnames.  Since the parameter reference gets mapped to a FUNC_TRUE condition with
the left side equal to the key and the right side equal to value, key and value were treated
as field names.

This change yanks the left and ride-side values of FUNC_TRUE conditions after the
parameters have been extracted.  The change also includes special handling in the
DataServiceTestController to avoid leaking any modified parameters, and to reset
them in between SQL execution.