[BACKLOG-5268] Alias and column order can be incorrect when using Group By and aggregate functions.
Formerly the Select values step in the generated transformation was
omitted if aggregates were present in the select field.  This was
assumed to be valid since fields and aggregate alias could be handled
in the Group By, but failed to handle cases where the aggregate field
was to the left of any of the GROUP BY columns.  Also failed to handle
aliases for the GROUP BY columns.