[BACKLOG-11677]Write query results to safe pipe
If the client DataInputStream is closed before the query is completely written, an IOExcepiton may be logged. This change makes the write stream defensive.

If the input stream is closed, the output stream will remain open. More data may be written to the output stream, but it will be ignored.